requirements.txt
numpy==1.18.1
oauthlib==3.1.0
Pillow==7.0.0
rsa==4.0
scipy==1.4.1
sh==1.12.14
SQLAlchemy==1.3.13
tensorflow==1.14.0
tensorflow-estimator==1.14.0
torch==1.4.0
torchvision==0.5.0
twython==3.7.0
schedule==0.6.0
